+++ geomstats/geometry/special_orthogonal3.py
