+++ tests/test_special_orthogonal.py
@@ -9,7 +9,7 @@ import tests.helper as helper
 import geomstats.backend as gs
 import geomstats.tests
 from geomstats.geometry.invariant_metric import InvariantMetric
-from geomstats.geometry.so3 import SpecialOrthogonal3
+from geomstats.geometry.special_orthogonal3 import SpecialOrthogonal3
 
 
 EPSILON = 1e-5
